+++ src/main/java/com/ecom/app/repository/OptionRepository.java
@@ -3,5 +3,5 @@ package com.ecom.app.repository;
 import com.ecom.app.entity.Option;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OptionRepository extends JpaRepository<Option, Integer> {
+public interface OptionRepository extends JpaRepository<Option, Long> {
 }
