+++ src/main/java/com/ecom/app/repository/PollRepository.java
@@ -3,5 +3,5 @@ package com.ecom.app.repository;
 import com.ecom.app.entity.Poll;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PollRepository extends JpaRepository<Poll, Integer> {
+public interface PollRepository extends JpaRepository<Poll, Long> {
 }
