+++ src/main/java/com/ecom/app/repository/ReturnRepository.java
@@ -3,5 +3,5 @@ package com.ecom.app.repository;
 import com.ecom.app.entity.Return;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReturnRepository extends JpaRepository<Return, Integer> {
+public interface ReturnRepository extends JpaRepository<Return, Long> {
 }
