+++ src/main/java/com/ecom/app/repository/ShippingRepository.java
@@ -3,5 +3,5 @@ package com.ecom.app.repository;
 import com.ecom.app.entity.Shipping;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShippingRepository extends JpaRepository<Shipping, Integer> {
+public interface ShippingRepository extends JpaRepository<Shipping, Long> {
 }
