+++ src/main/java/com/ecom/app/repository/UserDetailRepository.java
@@ -3,5 +3,5 @@ package com.ecom.app.repository;
 import com.ecom.app.entity.UserDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDetailRepository extends JpaRepository<UserDetail, Integer> {
+public interface UserDetailRepository extends JpaRepository<UserDetail, Long> {
 }
