+++ src/main/java/com/ecom/app/repository/VoteRepository.java
@@ -3,5 +3,5 @@ package com.ecom.app.repository;
 import com.ecom.app.entity.Vote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VoteRepository extends JpaRepository<Vote, Integer> {
+public interface VoteRepository extends JpaRepository<Vote, Long> {
 }
